Create numberofsteps.py
Given two arrays a and b. In each step, you can set a[i]=a[i]-b[i]  if a[i]>=b[i]. Determine the minimum number of steps that are required to make all a's equal.

Print the minimum number of steps that are required to make all a's equal. If it is not possible, then print -1.


Sample input

2
5 6
4 3

Sample output

-1

Sample Input
5
5 7 10 5 15
2 2 1 3 5
Sample Output
8