Create divisibilty.py
You are provided an array a of size n that contains non-negative integers. Your task is to determine whether the number that is formed by selecting the last digit of all the N numbers is divisible by 10.

Note: View the sample explanation section for more clarification.

Input format

First line: A single integer n denoting the size of array a 
Second line:  space-separated integers.
Output format

If the number is divisible by 10, then print yes. Otherwise, print no.
Explanation
Last digit of  is .
Last digit of  is .
Last digit of  is .
Last digit of  is .
Last digit of  is .
Therefore the number formed is 55514 which is not divisible by 10.